Move search tests to own file
